requirements.txt
beautifulsoup4==4.12.2
Flask==2.2.2
mecab_python3==1.0.7
requests==2.29.0
torch==2.0.1
transformers==4.33.1
pip==23.1.2
gunicorn==21.2.0

click==8.1.3
colorama==0.4.6
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
gdown==4.7.1